+++ packages/core/stellar/src/set-access-control.ts
@@ -1,9 +1,14 @@
 import type { BaseFunction, BaseTraitImplBlock, ContractBuilder } from './contract';
 
-export const accessOptions = [false, 'ownable'] as const;
+export const accessOptions = [false, 'ownable', 'roles'] as const;
 export const DEFAULT_ACCESS_CONTROL = 'ownable';
 
 export type Access = (typeof accessOptions)[number];
+export type AccessProps = {
+  useMacro: boolean;
+  caller?: string;
+  role?: string;
+};
 
 /**
  * Sets access control for the contract via constructor args.
@@ -15,12 +20,38 @@ export function setAccessControl(c: ContractBuilder, access: Access): void {
     case 'ownable': {
       if (!c.ownable) {
         c.ownable = true;
-        c.addUseClause('soroban_sdk', 'symbol_short');
-        c.addUseClause('soroban_sdk', 'Symbol');
-        c.addVariable({ name: 'OWNER', type: 'Symbol', value: `symbol_short!("OWNER")` });
+        c.addUseClause('soroban_sdk', 'Address');
+        c.addUseClause('stellar_ownable', 'self', { alias: 'ownable' });
+        c.addUseClause('stellar_ownable', 'Ownable');
+
+        const ownableTrait = {
+          traitName: 'Ownable',
+          structName: c.name,
+          tags: ['default_impl', 'contractimpl'],
+          section: 'Utils',
+        };
+        c.addTraitImplBlock(ownableTrait);
+
         c.addConstructorArgument({ name: 'owner', type: 'Address' });
-        c.addConstructorCode('e.storage().instance().set(&OWNER, &owner);');
+        c.addConstructorCode('ownable::set_owner(e, &owner);');
       }
+      break;
+    }
+    case 'roles': {
+      c.addUseClause('soroban_sdk', 'Address');
+      c.addUseClause('stellar_access_control', 'self', { alias: 'access_control' });
+      c.addUseClause('stellar_access_control', 'AccessControl');
+
+      const accessControltrait = {
+        traitName: 'AccessControl',
+        structName: c.name,
+        tags: ['default_impl', 'contractimpl'],
+        section: 'Utils',
+      };
+      c.addTraitImplBlock(accessControltrait);
+
+      c.addConstructorArgument({ name: 'admin', type: 'Address' });
+      c.addConstructorCode('access_control::set_admin(e, &admin);');
       break;
     }
     default: {
@@ -32,15 +63,13 @@ export function setAccessControl(c: ContractBuilder, access: Access): void {
 
 /**
  * Enables access control for the contract and restricts the given function with access control.
- *
- * If `caller` is provided, requires that the caller is the owner. Otherwise, requires that the owner is authorized.
  */
 export function requireAccessControl(
   c: ContractBuilder,
   trait: BaseTraitImplBlock | undefined,
   fn: BaseFunction,
   access: Access,
-  caller?: string,
+  accessProps: AccessProps = { useMacro: true },
 ): void {
   if (access === false) {
     access = DEFAULT_ACCESS_CONTROL;
@@ -49,19 +78,43 @@ export function requireAccessControl(
 
   switch (access) {
     case 'ownable': {
-      c.addUseClause('soroban_sdk', 'Address');
-      const getOwner = 'let owner: Address = e.storage().instance().get(&OWNER).expect("owner should be set");';
-      if (caller) {
-        c.addUseClause('soroban_sdk', 'panic_with_error');
-        c.addError('Unauthorized', 1); // TODO: Ensure there are no conflicts in error codes
-        c.addFunctionCodeBefore(
-          fn,
-          [getOwner, `if owner != ${caller} {`, `    panic_with_error!(e, ${c.name}Error::Unauthorized)`, '}'],
-          trait,
-        );
+      c.addUseClause('stellar_ownable_macro', 'only_owner');
+
+      if (accessProps.useMacro) {
+        c.addFunctionTag(fn, 'only_owner', trait);
       } else {
-        c.addFunctionCodeBefore(fn, [getOwner, 'owner.require_auth();'], trait);
+        c.addFunctionCodeBefore(fn, [`ownable::enforce_owner_auth(e);`], trait);
       }
+
+      break;
+    }
+    case 'roles': {
+      const { useMacro, caller, role } = accessProps;
+
+      if (caller && role) {
+        c.addUseClause('soroban_sdk', 'Symbol');
+        c.addConstructorArgument({ name: role, type: 'Address' });
+        c.addConstructorCode(`access_control::grant_role_no_auth(e, &admin, &${role}, &Symbol::new(e, "${role}"));`);
+
+        if (useMacro) {
+          c.addUseClause('stellar_access_control_macros', 'only_role');
+          c.addFunctionTag(fn, `only_role(${caller}, "${role}")`, trait);
+        } else {
+          c.addFunctionCodeBefore(
+            fn,
+            [`access_control::ensure_role(e, ${caller}, &Symbol::new(e, "${role}"));`],
+            trait,
+          );
+        }
+      } else {
+        if (useMacro) {
+          c.addUseClause('stellar_access_control_macros', 'only_admin');
+          c.addFunctionTag(fn, 'only_admin', trait);
+        } else {
+          c.addFunctionCodeBefore(fn, ['access_control::enforce_admin_auth(e);'], trait);
+        }
+      }
+
       break;
     }
     default: {